# peyda_docs/initial/OpenAPI.yaml
openapi: 3.0.3
info:
  title: سامانه پیدا - مدیریت گمشدگان مناسبتی
  description: |
    سامانه جامع مدیریت گمشدگان مناسبتی برای نیمه شعبان و سایر تجمعات انبوه.
    این API امکان ثبت گزارش‌های گمشده و پیدا شده، جستجو و تطبیق هوشمند،
    و مدیریت موکب‌ها را فراهم می‌کند.
  version: 1.1.0
  contact:
    name: تیم توسعه سامانه پیدا
    email: support@peyda.ir
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.peyda.ir/v1
    description: سرور تولید
  - url: https://staging-api.peyda.ir/v1
    description: سرور تست
paths:
...
  /auth/send-otp:
    post:
      tags:
        - احراز هویت
      summary: ارسال کد یکبار مصرف
      description: |
        <div dir="rtl">
        ارسال کد OTP به شماره موبایل کاربر از طریق پیام‌رسان‌های ایتا و واتساپ.
        در صورت شکست یا اتمام زمان، کاربر باید شماره موبایل را مجددا وارد کند.
        </div>
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - country_code
              properties:
                phone:
                  type: string
                  description: شماره موبایل کاربر
                  example: "9123456789"
                country_code:
                  type: string
                  description: کد کشور
                  example: "+98"
      responses:
        '200':
          description: کد OTP با موفقیت ارسال شد
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "کد تایید به شماره شما ارسال شد"
                  request_id:
                    type: string
                    description: شناسه درخواست OTP برای استفاده در تایید و ارسال مجدد
                    example: "req_1234567890"
                  expires_in:
                    type: integer
                    description: زمان انقضای کد به ثانیه
                    example: 300
                  max_attempts:
                    type: integer
                    description: حداکثر تعداد تلاش‌های مجاز برای ورود کد
                    example: 3
                  max_resends:
                    type: integer
                    description: حداکثر تعداد ارسال مجدد کد
                    example: 3
        '400':
          description: شماره موبایل نامعتبر
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_phone:
                  value:
                    status: 400
                    code: "INVALID_PHONE"
                    title: "شماره موبایل نامعتبر"
                    message: "شماره موبایل وارد شده صحیح نیست."
        '429':
          description: درخواست‌های بیش از حد مجاز
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                too_many_requests:
                  value:
                    status: 429
                    code: "TOO_MANY_REQUESTS"
                    title: "درخواست‌های زیاد"
                    message: "تعداد درخواست‌های شما بیش از حد مجاز است. لطفاً بعداً تلاش کنید."
  /auth/verify-otp:
    post:
      tags:
        - احراز هویت
      summary: تایید کد یکبار مصرف و ورود
      description: |
        <div dir="rtl">
        تایید کد OTP و دریافت توکن دسترسی.
        کاربر تا ۳ بار فرصت ورود کد صحیح دارد.
        </div>
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - otp
                - request_id
              properties:
                otp:
                  type: string
                  description: کد یکبار مصرف
                  example: "123456"
                request_id:
                  type: string
                  description: شناسه درخواست OTP دریافتی از endpoint ارسال کد
                  example: "req_1234567890"
      responses:
        '200':
          description: ورود موفقیت‌آمیز
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    description: توکن JWT برای احراز هویت
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: کد OTP نامعتبر یا منقضی شده
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_otp:
                  value:
                    status: 400
                    code: "INVALID_OTP"
                    title: "کد تایید نامعتبر"
                    message: "کد تایید وارد شده صحیح نیست."
                otp_expired:
                  value:
                    status: 400
                    code: "OTP_EXPIRED"
                    title: "کد تایید منقضی شده"
                    message: "کد تایید وارد شده منقضی شده است. لطفاً مجدداً درخواست کد جدید بدهید."
                invalid_request_id:
                  value:
                    status: 400
                    code: "INVALID_REQUEST_ID"
                    title: "شناسه درخواست نامعتبر"
                    message: "شناسه درخواست OTP معتبر نیست."
        '403':
          description: تعداد تلاش‌های بیش از حد مجاز
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                max_attempts:
                  value:
                    status: 403
                    code: "MAX_ATTEMPTS_REACHED"
                    title: "حداکثر تلاش"
                    message: "تعداد تلاش‌های شما برای ورود کد به حداکثر رسیده است. لطفاً کد جدید درخواست دهید."
  /auth/resend-otp:
    post:
      tags:
        - احراز هویت
      summary: ارسال مجدد کد یکبار مصرف
      description: |
        <div dir="rtl">
        ارسال مجدد کد OTP برای درخواست موجود.
        فقط در صورتی امکان‌پذیر است که تعداد ارسال مجدد به حداکثر نرسیده باشد.
        </div>
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - request_id
              properties:
                request_id:
                  type: string
                  description: شناسه درخواست OTP که می‌خواهید مجددا ارسال شود
                  example: "req_1234567890"
      responses:
        '200':
          description: کد OTP با موفقیت مجددا ارسال شد
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "کد تایید مجددا ارسال شد"
                  request_id:
                    type: string
                    description: شناسه درخواست OTP (همان شناسه قبلی)
                    example: "req_1234567890"
                  expires_in:
                    type: integer
                    description: زمان انقضای کد جدید به ثانیه
                    example: 300
                  remaining_resends:
                    type: integer
                    description: تعداد باقی‌مانده ارسال مجدد
                    example: 2
        '400':
          description: شناسه درخواست نامعتبر یا منقضی شده
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_request_id:
                  value:
                    status: 400
                    code: "INVALID_REQUEST_ID"
                    title: "شناسه درخواست نامعتبر"
                    message: "شناسه درخواست OTP معتبر نیست یا منقضی شده است."
        '429':
          description: حداکثر تعداد ارسال مجدد استفاده شده است

  # مدیریت مدیا
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: شناسه کاربر
          example: "user_123"
        phone:
          type: string
          description: شماره موبایل کاربر
          example: "9123456789"
        role:
          type: string
          enum: [user, mawkab_owner, admin]
          description: نقش کاربر
          example: "mawkab_owner"
        is_verified:
          type: boolean
          description: آیا موکب کاربر تایید شده است
          example: true
        created_at:
          type: integer
          description: تاریخ ثبت‌نام کاربر (ثانیه از epoch)
          example: 1708008600
...
    Error:
      type: object
      properties:
        status:
          type: integer
          description: کد وضعیت HTTP
          example: 400
        code:
          type: string
          description: کد خطای یکتا برای برنامه
          example: "OTP_EXPIRED"
        title:
          type: string
          description: عنوان کوتاه خطا
          example: "کد تایید منقضی شده"
        message:
          type: string
          description: پیام کامل خطا برای کاربر
          example: "کد تایید وارد شده منقضی شده است. لطفاً مجدداً درخواست کد جدید بدهید."